Remove code visibility
This feature was never used and is not a priority
