+++ .env.development
@@ -29,7 +29,6 @@ FEATURE_REPORTING_CONFIGURATIONS=true
 FEATURE_ANALYTICS=true
 FEATURE_SUPPORT=true
 FEATURE_SAML_CONFIGURATION=true
-FEATURE_CODE_VISIBILITY=true
 FEATURE_EXTERNAL_LMS_CONFIGURATION=true
 FEATURE_BULK_ENROLLMENT=true
 HOTJAR_APP_ID=''
