+++ src/components/CouponDetails/CouponBulkActions.jsx
@@ -9,7 +9,7 @@ const CouponBulkActions = ({
   value, options, disabled, onChange, handleBulkAction,
 }) => (
   <div className="bulk-actions col-12 col-md-4 text-md-right mt-3 m-md-0 d-flex justify-content-end">
-    <Form.Group className="mt-1 flex-grow-1" controlId="visibility" name="bulk-actions">
+    <Form.Group className="mt-1 flex-grow-1" controlId="bulkActions" name="bulk-actions">
       <Form.Control
         className="float-right w-50"
         floatingLabel="Bulk action"
