+++ src/components/CouponDetails/constants.js
@@ -7,51 +7,37 @@ export const COUPON_FILTER_TYPES = {
   redeemed: 'redeemed',
 };
 
-export const COUPON_VISIBILITY_OPTIONS = {
-  both: 'both',
-  public: 'public',
-  private: 'private',
+export const COUPON_FILTER_OPTIONS = {
+  unassigned: 'Unassigned',
+  unredeemed: 'Unredeemed',
+  partiallyRedeemed: 'Partially redeemed',
+  redeemed: 'Redeemed',
 };
 
 export const ACTION_LABELS = {
   remind: 'Remind',
   assign: 'Assign',
   revoke: 'Revoke',
-  makePublic: 'Make public',
-  makePrivate: 'Make private',
 };
 
 export const ACTION_TYPES = {
   remind: 'remind',
   revoke: 'revoke',
   assign: 'assign',
-  makePublic: 'make_public',
-  makePrivate: 'make_private',
 };
 
 export const FILTER_OPTIONS = [{
-  label: 'Unassigned',
+  label: COUPON_FILTER_OPTIONS.unassigned,
   value: COUPON_FILTER_TYPES.unassigned,
 }, {
-  label: 'Unredeemed',
+  label: COUPON_FILTER_OPTIONS.unredeemed,
   value: COUPON_FILTER_TYPES.unredeemed,
 }, {
-  label: 'Partially Redeemed',
+  label: COUPON_FILTER_OPTIONS.partiallyRedeemed,
   value: COUPON_FILTER_TYPES.partiallyRedeemed,
 }, {
-  label: 'Redeemed',
+  label: COUPON_FILTER_OPTIONS.redeemed,
   value: COUPON_FILTER_TYPES.redeemed,
-}];
-
-export const VISIBILITY_OPTIONS = [{
-  label: 'Both',
-  value: 'both',
-}, {
-  label: 'Public',
-  value: 'public',
-}, {
-  label: 'Private',
-  value: 'private',
 }];
 
 export const BULK_ACTION_SELECT_OPTIONS = [{
@@ -82,10 +68,6 @@ export const COLUMNS = {
     label: 'Actions',
     key: 'actions',
   },
-  visibility: {
-    label: 'Visibility',
-    key: 'is_public',
-  },
   lastReminderDate: {
     label: 'Last reminder date',
     key: 'last_reminder_date',
@@ -102,4 +84,38 @@ export const COLUMNS = {
     label: 'Redeemed by',
     key: 'assigned_to',
   },
+};
+
+const COMMON_COLUMNS = [
+  COLUMNS.redemptions,
+  COLUMNS.code,
+];
+
+const REDEMTION_COLUMNS = [
+  ...COMMON_COLUMNS,
+  COLUMNS.assignmentDate,
+  COLUMNS.lastReminderDate,
+  COLUMNS.actions,
+];
+
+export const DEFAULT_TABLE_COLUMNS = {
+  [COUPON_FILTER_TYPES.unassigned]: [
+    ...COMMON_COLUMNS,
+    COLUMNS.assignmentsRemaining,
+    COLUMNS.actions,
+  ],
+  [COUPON_FILTER_TYPES.unredeemed]: [
+    COLUMNS.assignedTo,
+    ...REDEMTION_COLUMNS,
+  ],
+  [COUPON_FILTER_TYPES.partiallyRedeemed]: [
+    COLUMNS.assignedTo,
+    ...REDEMTION_COLUMNS,
+  ],
+  [COUPON_FILTER_TYPES.redeemed]: [
+    COLUMNS.redeemedBy,
+    ...COMMON_COLUMNS,
+    COLUMNS.assignmentDate,
+    COLUMNS.lastReminderDate,
+  ],
 };
