+++ src/config/index.js
@@ -45,7 +45,6 @@ const features = {
   ANALYTICS: process.env.FEATURE_ANALYTICS,
   SAML_CONFIGURATION: process.env.FEATURE_SAML_CONFIGURATION,
   SUPPORT: process.env.FEATURE_SUPPORT,
-  CODE_VISIBILITY: process.env.FEATURE_CODE_VISIBILITY,
   EXTERNAL_LMS_CONFIGURATION: process.env.FEATURE_EXTERNAL_LMS_CONFIGURATION,
   ANALYTICS_API_V1: process.env.FEATURE_ANALYTICS_API_V1 || hasFeatureFlagEnabled(FEATURE_ANALYTICS_API_V1),
 };
