--- src/data/constants/codeVisibility.js
@@ -1,9 +1,0 @@
-const CODE_VISIBILITY_REQUEST = 'CODE_VISIBILITY_REQUEST';
-const CODE_VISIBILITY_SUCCESS = 'CODE_VISIBILITY_SUCCESS';
-const CODE_VISIBILITY_FAILURE = 'CODE_VISIBILITY_FAILURE';
-
-export {
-  CODE_VISIBILITY_REQUEST,
-  CODE_VISIBILITY_SUCCESS,
-  CODE_VISIBILITY_FAILURE,
-};
